+++ fltk-sys/cfltk/cfl_gl.h
@@ -5,3 +5,8 @@
 #else
 #include <GL/glu.h>
 #endif
+
+void gl_start(void);
+
+void gl_finish(void);
+