+++ fltk-sys/src/button.rs
@@ -234,6 +234,21 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_set_handle_data(self_: *mut Fl_Widget, data: *mut ::std::os::raw::c_void);
 }
+extern "C" {
+    pub fn Fl_Widget_damage(self_: *const Fl_Widget) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Widget_set_damage(self_: *mut Fl_Widget, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Widget_clear_damage(self_: *mut Fl_Widget);
+}
+extern "C" {
+    pub fn Fl_Widget_as_window(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 #[repr(C)]
 #[derive(Debug, Copy, Clone)]
 pub struct Fl_Button {
@@ -449,6 +464,21 @@ extern "C" {
 }
 extern "C" {
     pub fn Fl_Button_set_handle_data(self_: *mut Fl_Button, data: *mut ::std::os::raw::c_void);
+}
+extern "C" {
+    pub fn Fl_Button_damage(self_: *const Fl_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Button_set_damage(self_: *mut Fl_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Button_clear_damage(self_: *mut Fl_Button);
+}
+extern "C" {
+    pub fn Fl_Button_as_window(self_: *mut Fl_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Button_as_group(self_: *mut Fl_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Button_shortcut(self_: *const Fl_Button) -> ::std::os::raw::c_int;
@@ -704,6 +734,21 @@ extern "C" {
         self_: *mut Fl_Check_Button,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Check_Button_damage(self_: *const Fl_Check_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Check_Button_set_damage(self_: *mut Fl_Check_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Check_Button_clear_damage(self_: *mut Fl_Check_Button);
+}
+extern "C" {
+    pub fn Fl_Check_Button_as_window(self_: *mut Fl_Check_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Check_Button_as_group(self_: *mut Fl_Check_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Check_Button_is_checked(arg1: *mut Fl_Check_Button) -> ::std::os::raw::c_int;
@@ -968,6 +1013,21 @@ extern "C" {
         self_: *mut Fl_Radio_Button,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Radio_Button_damage(self_: *const Fl_Radio_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Radio_Button_set_damage(self_: *mut Fl_Radio_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Radio_Button_clear_damage(self_: *mut Fl_Radio_Button);
+}
+extern "C" {
+    pub fn Fl_Radio_Button_as_window(self_: *mut Fl_Radio_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Radio_Button_as_group(self_: *mut Fl_Radio_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Radio_Button_is_toggled(arg1: *mut Fl_Radio_Button) -> ::std::os::raw::c_int;
@@ -1246,6 +1306,21 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Toggle_Button_damage(self_: *const Fl_Toggle_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Toggle_Button_set_damage(self_: *mut Fl_Toggle_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Toggle_Button_clear_damage(self_: *mut Fl_Toggle_Button);
+}
+extern "C" {
+    pub fn Fl_Toggle_Button_as_window(self_: *mut Fl_Toggle_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Toggle_Button_as_group(self_: *mut Fl_Toggle_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Toggle_Button_is_toggled(arg1: *mut Fl_Toggle_Button) -> ::std::os::raw::c_int;
 }
 extern "C" {
@@ -1508,6 +1583,21 @@ extern "C" {
         self_: *mut Fl_Round_Button,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Round_Button_damage(self_: *const Fl_Round_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Round_Button_set_damage(self_: *mut Fl_Round_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Round_Button_clear_damage(self_: *mut Fl_Round_Button);
+}
+extern "C" {
+    pub fn Fl_Round_Button_as_window(self_: *mut Fl_Round_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Round_Button_as_group(self_: *mut Fl_Round_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Round_Button_is_toggled(arg1: *mut Fl_Round_Button) -> ::std::os::raw::c_int;
@@ -1836,6 +1926,30 @@ extern "C" {
         self_: *mut Fl_Radio_Round_Button,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_damage(
+        self_: *const Fl_Radio_Round_Button,
+    ) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_set_damage(
+        self_: *mut Fl_Radio_Round_Button,
+        flag: ::std::os::raw::c_uchar,
+    );
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_clear_damage(self_: *mut Fl_Radio_Round_Button);
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_as_window(
+        self_: *mut Fl_Radio_Round_Button,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Radio_Round_Button_as_group(
+        self_: *mut Fl_Radio_Round_Button,
+    ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Radio_Round_Button_is_toggled(
@@ -2176,6 +2290,30 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Radio_Light_Button_damage(
+        self_: *const Fl_Radio_Light_Button,
+    ) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Radio_Light_Button_set_damage(
+        self_: *mut Fl_Radio_Light_Button,
+        flag: ::std::os::raw::c_uchar,
+    );
+}
+extern "C" {
+    pub fn Fl_Radio_Light_Button_clear_damage(self_: *mut Fl_Radio_Light_Button);
+}
+extern "C" {
+    pub fn Fl_Radio_Light_Button_as_window(
+        self_: *mut Fl_Radio_Light_Button,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Radio_Light_Button_as_group(
+        self_: *mut Fl_Radio_Light_Button,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Radio_Light_Button_is_toggled(
         arg1: *mut Fl_Radio_Light_Button,
     ) -> ::std::os::raw::c_int;
@@ -2448,6 +2586,21 @@ extern "C" {
         self_: *mut Fl_Light_Button,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Light_Button_damage(self_: *const Fl_Light_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Light_Button_set_damage(self_: *mut Fl_Light_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Light_Button_clear_damage(self_: *mut Fl_Light_Button);
+}
+extern "C" {
+    pub fn Fl_Light_Button_as_window(self_: *mut Fl_Light_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Light_Button_as_group(self_: *mut Fl_Light_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Light_Button_is_on(arg1: *mut Fl_Light_Button) -> ::std::os::raw::c_int;
@@ -2726,6 +2879,21 @@ extern "C" {
     );
 }
 extern "C" {
+    pub fn Fl_Repeat_Button_damage(self_: *const Fl_Repeat_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Repeat_Button_set_damage(self_: *mut Fl_Repeat_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Repeat_Button_clear_damage(self_: *mut Fl_Repeat_Button);
+}
+extern "C" {
+    pub fn Fl_Repeat_Button_as_window(self_: *mut Fl_Repeat_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Repeat_Button_as_group(self_: *mut Fl_Repeat_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
     pub fn Fl_Repeat_Button_shortcut(self_: *const Fl_Repeat_Button) -> ::std::os::raw::c_int;
 }
 extern "C" {
@@ -2994,6 +3162,21 @@ extern "C" {
         self_: *mut Fl_Return_Button,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Return_Button_damage(self_: *const Fl_Return_Button) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Return_Button_set_damage(self_: *mut Fl_Return_Button, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Return_Button_clear_damage(self_: *mut Fl_Return_Button);
+}
+extern "C" {
+    pub fn Fl_Return_Button_as_window(self_: *mut Fl_Return_Button) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Return_Button_as_group(self_: *mut Fl_Return_Button) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Return_Button_shortcut(self_: *const Fl_Return_Button) -> ::std::os::raw::c_int;