+++ fltk-sys/src/text.rs
@@ -234,6 +234,21 @@ extern "C" {
 extern "C" {
     pub fn Fl_Widget_set_handle_data(self_: *mut Fl_Widget, data: *mut ::std::os::raw::c_void);
 }
+extern "C" {
+    pub fn Fl_Widget_damage(self_: *const Fl_Widget) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Widget_set_damage(self_: *mut Fl_Widget, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Widget_clear_damage(self_: *mut Fl_Widget);
+}
+extern "C" {
+    pub fn Fl_Widget_as_window(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Widget_as_group(self_: *mut Fl_Widget) -> *mut ::std::os::raw::c_void;
+}
 pub type Fl_Text_Modify_Cb = ::std::option::Option<
     unsafe extern "C" fn(
         pos: ::std::os::raw::c_int,
@@ -674,6 +689,21 @@ extern "C" {
         self_: *mut Fl_Text_Display,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Text_Display_damage(self_: *const Fl_Text_Display) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Text_Display_set_damage(self_: *mut Fl_Text_Display, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Text_Display_clear_damage(self_: *mut Fl_Text_Display);
+}
+extern "C" {
+    pub fn Fl_Text_Display_as_window(self_: *mut Fl_Text_Display) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Text_Display_as_group(self_: *mut Fl_Text_Display) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Text_Display_init(arg1: *mut Fl_Text_Display);
@@ -1159,6 +1189,21 @@ extern "C" {
         self_: *mut Fl_Text_Editor,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Text_Editor_damage(self_: *const Fl_Text_Editor) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Text_Editor_set_damage(self_: *mut Fl_Text_Editor, flag: ::std::os::raw::c_uchar);
+}
+extern "C" {
+    pub fn Fl_Text_Editor_clear_damage(self_: *mut Fl_Text_Editor);
+}
+extern "C" {
+    pub fn Fl_Text_Editor_as_window(self_: *mut Fl_Text_Editor) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Text_Editor_as_group(self_: *mut Fl_Text_Editor) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Text_Editor_init(arg1: *mut Fl_Text_Editor);
@@ -1697,6 +1742,28 @@ extern "C" {
         self_: *mut Fl_Simple_Terminal,
         data: *mut ::std::os::raw::c_void,
     );
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_damage(self_: *const Fl_Simple_Terminal) -> ::std::os::raw::c_uchar;
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_set_damage(
+        self_: *mut Fl_Simple_Terminal,
+        flag: ::std::os::raw::c_uchar,
+    );
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_clear_damage(self_: *mut Fl_Simple_Terminal);
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_as_window(
+        self_: *mut Fl_Simple_Terminal,
+    ) -> *mut ::std::os::raw::c_void;
+}
+extern "C" {
+    pub fn Fl_Simple_Terminal_as_group(
+        self_: *mut Fl_Simple_Terminal,
+    ) -> *mut ::std::os::raw::c_void;
 }
 extern "C" {
     pub fn Fl_Simple_Terminal_init(arg1: *mut Fl_Simple_Terminal);